Rename cnn_framework to dde
cnn_framework was the name left over from when this was still a part
of RMG.
