+++ Makefile
@@ -9,4 +9,4 @@ ifneq ($(OS),Windows_NT)
 	mkdir -p testing/coverage
 	rm -rf testing/coverage/*
 endif
-	nosetests --nocapture --nologcapture --all-modules -A 'not helper' --verbose --with-coverage --cover-inclusive --cover-package=cnn_framework --cover-erase --cover-html --cover-html-dir=testing/coverage --exe cnn_framework
+	nosetests --nocapture --nologcapture --all-modules -A 'not helper' --verbose --with-coverage --cover-inclusive --cover-package=dde --cover-erase --cover-html --cover-html-dir=testing/coverage --exe dde
