+++ dde/predictor.py
@@ -1,16 +1,16 @@
 #!/usr/bin/env python
 # -*- coding:utf-8 -*-
 
-from cnn_framework.cnn_model import build_model, train_model, reset_model, save_model, write_loss_report
-from cnn_framework.input import read_input_file
-from cnn_framework.molecule_tensor import get_molecule_tensor, pad_molecule_tensor
+from dde.cnn_model import build_model, train_model, reset_model, save_model, write_loss_report
+from dde.input import read_input_file
+from dde.molecule_tensor import get_molecule_tensor, pad_molecule_tensor
 import os
 import shutil
 import rmgpy
 import numpy as np
-from cnn_framework.data import (prepare_data_one_fold, prepare_folded_data_from_multiple_datasets,
-                                prepare_full_train_data_from_multiple_datasets, split_inner_val_from_train_data,
-                                prepare_folded_data_from_file, prepare_full_train_data_from_file)
+from dde.data import (prepare_data_one_fold, prepare_folded_data_from_multiple_datasets,
+                      prepare_full_train_data_from_multiple_datasets, split_inner_val_from_train_data,
+                      prepare_folded_data_from_file, prepare_full_train_data_from_file)
 import logging
 from keras.callbacks import EarlyStopping
 import json
