+++ scripts/evaluate_cnn.py
@@ -8,8 +8,8 @@ import numpy as np
 import pandas as pd
 from tqdm import tqdm
 
-from cnn_framework.data import get_db_mols
-from cnn_framework.predictor import Predictor
+from dde.data import get_db_mols
+from dde.predictor import Predictor
 
 
 def parse_command_line_arguments():
